Fixed issues with modify upload
