+++ mmeds/tools/__init__.py
@@ -1,0 +1,6 @@
+__author__ = "The Clemente Lab"
+__copyright__ = "Copyright (c) 2016 The Clemente Lab"
+__credits__ = ["David S. Wallach", "Jose C. Clemente"]
+__license__ = "GPL"
+__maintainer__ = "David S. Wallach"
+__email__ = "d.s.t.wallach@gmail.com"
